+++ blocks-src/jb-jobs-dashboard/block.json
