+++ blocks-src/jb-recent-jobs/block.json
@@ -1,8 +1,8 @@
 {
   "apiVersion": 2,
-  "name": "jb-block/jb-jobs-list",
-  "title": "Jobs list",
-  "description": "Displaying jobs list",
+  "name": "jb-block/jb-recent-jobs",
+  "title": "Recent Jobs",
+  "description": "Displaying recent jobs",
   "icon": "editor-ul",
   "category": "jb-blocks",
   "editorScript": "file:./src/index.js",
