+++ includes/blocks/blocks-manifest.php
@@ -1,0 +1,64 @@
+<?php
+// This file is generated. Do not modify it manually.
+return array(
+	'jb-job' => array(
+		'apiVersion' => 2,
+		'name' => 'jb-block/jb-job',
+		'title' => 'Single job',
+		'description' => 'Displaying a single job',
+		'icon' => 'text',
+		'category' => 'jb-blocks',
+		'editorScript' => 'file:./src/index.js',
+		'textdomain' => 'jobboardwp'
+	),
+	'jb-job-post' => array(
+		'apiVersion' => 2,
+		'title' => 'Post job',
+		'description' => 'Displaying jobs posting form',
+		'name' => 'jb-block/jb-job-post',
+		'category' => 'jb-blocks',
+		'icon' => 'forms',
+		'editorScript' => 'file:./src/index.js',
+		'textdomain' => 'jobboardwp'
+	),
+	'jb-jobs-categories-list' => array(
+		'apiVersion' => 2,
+		'title' => 'Jobs categories list',
+		'description' => 'Displaying jobs categories list',
+		'name' => 'jb-block/jb-jobs-categories-list',
+		'category' => 'jb-blocks',
+		'icon' => 'editor-ul',
+		'editorScript' => 'file:./src/index.js',
+		'textdomain' => 'jobboardwp'
+	),
+	'jb-jobs-dashboard' => array(
+		'apiVersion' => 2,
+		'title' => 'Jobs dashboard',
+		'description' => 'Displaying jobs dashboard',
+		'name' => 'jb-block/jb-jobs-dashboard',
+		'category' => 'jb-blocks',
+		'icon' => 'dashboard',
+		'editorScript' => 'file:./src/index.js',
+		'textdomain' => 'jobboardwp'
+	),
+	'jb-jobs-list' => array(
+		'apiVersion' => 2,
+		'name' => 'jb-block/jb-jobs-list',
+		'title' => 'Jobs list',
+		'description' => 'Displaying jobs list',
+		'icon' => 'editor-ul',
+		'category' => 'jb-blocks',
+		'editorScript' => 'file:./src/index.js',
+		'textdomain' => 'jobboardwp'
+	),
+	'jb-recent-jobs' => array(
+		'apiVersion' => 2,
+		'name' => 'jb-block/jb-recent-jobs',
+		'title' => 'Recent Jobs',
+		'description' => 'Displaying recent jobs',
+		'icon' => 'editor-ul',
+		'category' => 'jb-blocks',
+		'editorScript' => 'file:./src/index.js',
+		'textdomain' => 'jobboardwp'
+	)
+);
