+++ includes/blocks/jb-job-post/src/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-element', 'wp-server-side-render'), 'version' => '386bbdf966d94d8d34bf');
