+++ includes/blocks/jb-job-post/src/index.js
@@ -1,46 +1,1 @@
-import { registerBlockType } from '@wordpress/blocks';
-import ServerSideRender from '@wordpress/server-side-render';
-import { useBlockProps } from '@wordpress/block-editor';
-import { useEffect } from '@wordpress/element';
-
-registerBlockType('jb-block/jb-job-post', {
-	edit: function () {
-		useEffect(() => {
-			document.querySelectorAll('#jb-job-preview, #jb-job-draft, #jb_company_logo_plupload')
-			.forEach(element => {
-				element.setAttribute('disabled', 'disabled');
-			});
-		}, []);
-
-		const blockProps = useBlockProps();
-		return (
-			<div {...blockProps}>
-				<ServerSideRender block="jb-block/jb-job-post" />
-			</div>
-		);
-	},
-	save: () => null
-});
-
-jQuery(window).on('load', function () {
-	const observer = new MutationObserver((mutations) => {
-		mutations.forEach((mutation) => {
-			jQuery(mutation.addedNodes)
-			.find('.jb-job-submission-form-wrapper')
-			.each(function () {
-				const wrapper = document.querySelector('.jb-job-submission-form-wrapper');
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-		});
-	});
-
-	observer.observe(document, {attributes: false, childList: true, characterData: false, subtree:true});
-});
+(()=>{"use strict";var e={n:o=>{var t=o&&o.__esModule?()=>o.default:()=>o;return e.d(t,{a:t}),t},d:(o,t)=>{for(var r in t)e.o(t,r)&&!e.o(o,r)&&Object.defineProperty(o,r,{enumerable:!0,get:t[r]})},o:(e,o)=>Object.prototype.hasOwnProperty.call(e,o)};const o=window.wp.blocks,t=window.wp.serverSideRender;var r=e.n(t);const n=window.wp.blockEditor,s=window.wp.element,c=window.ReactJSXRuntime;(0,o.registerBlockType)("jb-block/jb-job-post",{edit:function(){(0,s.useEffect)((()=>{document.querySelectorAll("#jb-job-preview, #jb-job-draft, #jb_company_logo_plupload").forEach((e=>{e.setAttribute("disabled","disabled")}))}),[]);const e=(0,n.useBlockProps)();return(0,c.jsx)("div",{...e,children:(0,c.jsx)(r(),{block:"jb-block/jb-job-post"})})},save:()=>null}),jQuery(window).on("load",(function(){new MutationObserver((e=>{e.forEach((e=>{jQuery(e.addedNodes).find(".jb-job-submission-form-wrapper").each((function(){const e=document.querySelector(".jb-job-submission-form-wrapper");e&&e.addEventListener("click",(o=>{o.target!==e&&(o.preventDefault(),o.stopPropagation())}))}))}))})).observe(document,{attributes:!1,childList:!0,characterData:!1,subtree:!0})}))})();
