+++ includes/blocks/jb-job/src/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-data', 'wp-server-side-render'), 'version' => 'ecd0a05f6fdd68e8b9f8');
