+++ includes/blocks/jb-job/src/index.js
@@ -1,80 +1,1 @@
-import { useSelect } from '@wordpress/data';
-import { PanelBody, SelectControl, Spinner } from '@wordpress/components';
-import { InspectorControls, useBlockProps } from '@wordpress/block-editor';
-import ServerSideRender from '@wordpress/server-side-render';
-import { registerBlockType } from "@wordpress/blocks";
-
-registerBlockType('jb-block/jb-job', {
-	edit: function (props) {
-		const { attributes, setAttributes } = props;
-		const { job_id } = attributes;
-		const blockProps = useBlockProps();
-
-		const posts = useSelect((select) => {
-			return select('core').getEntityRecords('postType', 'jb-job', {
-				per_page: -1,
-				_fields: ['id', 'title']
-			});
-		}, []);
-
-		if (!posts) {
-			return (
-				<p>
-					<Spinner />
-					{wp.i18n.__('Loading...', 'jobboardwp')}
-				</p>
-			);
-		}
-
-		if (posts.length === 0) {
-			return <p>{wp.i18n.__('Jobs not found', 'jobboardwp')}</p>;
-		}
-
-		const get_post = [{ label: '', value: '' }].concat(
-			posts.map((post) => ({
-				label: post.title.rendered,
-				value: post.id
-			}))
-		);
-
-		return (
-			<div {...blockProps}>
-				<ServerSideRender block="jb-block/jb-job" attributes={attributes} />
-				<InspectorControls>
-					<PanelBody title={wp.i18n.__('Job', 'jobboardwp')}>
-						<SelectControl
-							label={wp.i18n.__('Job', 'jobboardwp')}
-							className="jb_select_job"
-							value={job_id}
-							options={get_post}
-							onChange={(value) => setAttributes({ job_id: value })}
-						/>
-					</PanelBody>
-				</InspectorControls>
-			</div>
-		);
-	},
-
-	save: () => null
-});
-
-jQuery(window).on( 'load', function($) {
-	var observer = new MutationObserver(function(mutations) {
-		mutations.forEach(function(mutation) {
-			jQuery(mutation.addedNodes).find('.jb-single-job-wrapper').each(function() {
-				const wrapper = document.querySelector('.jb-single-job-wrapper');
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-		});
-	});
-
-	observer.observe(document, {attributes: false, childList: true, characterData: false, subtree:true});
-});
+(()=>{"use strict";var e={n:o=>{var t=o&&o.__esModule?()=>o.default:()=>o;return e.d(t,{a:t}),t},d:(o,t)=>{for(var n in t)e.o(t,n)&&!e.o(o,n)&&Object.defineProperty(o,n,{enumerable:!0,get:t[n]})},o:(e,o)=>Object.prototype.hasOwnProperty.call(e,o)};const o=window.wp.data,t=window.wp.components,n=window.wp.blockEditor,r=window.wp.serverSideRender;var i=e.n(r);const s=window.wp.blocks,a=window.ReactJSXRuntime;(0,s.registerBlockType)("jb-block/jb-job",{edit:function(e){const{attributes:r,setAttributes:s}=e,{job_id:d}=r,c=(0,n.useBlockProps)(),l=(0,o.useSelect)((e=>e("core").getEntityRecords("postType","jb-job",{per_page:-1,_fields:["id","title"]})),[]);if(!l)return(0,a.jsxs)("p",{children:[(0,a.jsx)(t.Spinner,{}),wp.i18n.__("Loading...","jobboardwp")]});if(0===l.length)return(0,a.jsx)("p",{children:wp.i18n.__("Jobs not found","jobboardwp")});const b=[{label:"",value:""}].concat(l.map((e=>({label:e.title.rendered,value:e.id}))));return(0,a.jsxs)("div",{...c,children:[(0,a.jsx)(i(),{block:"jb-block/jb-job",attributes:r}),(0,a.jsx)(n.InspectorControls,{children:(0,a.jsx)(t.PanelBody,{title:wp.i18n.__("Job","jobboardwp"),children:(0,a.jsx)(t.SelectControl,{label:wp.i18n.__("Job","jobboardwp"),className:"jb_select_job",value:d,options:b,onChange:e=>s({job_id:e})})})})]})},save:()=>null}),jQuery(window).on("load",(function(e){new MutationObserver((function(e){e.forEach((function(e){jQuery(e.addedNodes).find(".jb-single-job-wrapper").each((function(){const e=document.querySelector(".jb-single-job-wrapper");e&&e.addEventListener("click",(o=>{o.target!==e&&(o.preventDefault(),o.stopPropagation())}))}))}))})).observe(document,{attributes:!1,childList:!0,characterData:!1,subtree:!0})}))})();
