+++ includes/blocks/jb-jobs-categories-list/src/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-server-side-render'), 'version' => '8967b086c99cfc8aec5e');
