+++ includes/blocks/jb-jobs-categories-list/src/index.js
@@ -1,58 +1,1 @@
-jQuery(window).on( 'load', function($) {
-	var observer = new MutationObserver(function(mutations) {
-		mutations.forEach(function(mutation) {
-
-			jQuery(mutation.addedNodes).find('.jb-job-categories').each(function() {
-				wp.JB.job_categories_list.objects.wrapper = jQuery('.jb-job-categories');
-				if ( wp.JB.job_categories_list.objects.wrapper.length ) {
-					wp.JB.job_categories_list.ajax();
-				}
-				const wrapper = document.querySelector('.jb-job-categories');
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-
-			jQuery(mutation.addedNodes).find('.jb').each(function() {
-				jb_responsive();
-				const wrapper = document.querySelector('.jb-job-categories');
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-		});
-	});
-
-	observer.observe(document, {attributes: false, childList: true, characterData: false, subtree:true});
-});
-
-import { registerBlockType } from '@wordpress/blocks';
-import ServerSideRender from '@wordpress/server-side-render';
-import { useBlockProps } from '@wordpress/block-editor';
-
-registerBlockType('jb-block/jb-jobs-categories-list', {
-	edit: function(props) {
-		const blockProps = useBlockProps();
-
-		return (
-			<div {...blockProps}>
-				<ServerSideRender block="jb-block/jb-jobs-categories-list" />
-			</div>
-		);
-	},
-	save: function() {
-		return null;
-	}
-});
+(()=>{"use strict";var e={n:t=>{var o=t&&t.__esModule?()=>t.default:()=>t;return e.d(o,{a:o}),o},d:(t,o)=>{for(var r in o)e.o(o,r)&&!e.o(t,r)&&Object.defineProperty(t,r,{enumerable:!0,get:o[r]})},o:(e,t)=>Object.prototype.hasOwnProperty.call(e,t)};const t=window.wp.blocks,o=window.wp.serverSideRender;var r=e.n(o);const n=window.wp.blockEditor,c=window.ReactJSXRuntime;jQuery(window).on("load",(function(e){new MutationObserver((function(e){e.forEach((function(e){jQuery(e.addedNodes).find(".jb-job-categories").each((function(){wp.JB.job_categories_list.objects.wrapper=jQuery(".jb-job-categories"),wp.JB.job_categories_list.objects.wrapper.length&&wp.JB.job_categories_list.ajax();const e=document.querySelector(".jb-job-categories");e&&e.addEventListener("click",(t=>{t.target!==e&&(t.preventDefault(),t.stopPropagation())}))})),jQuery(e.addedNodes).find(".jb").each((function(){jb_responsive();const e=document.querySelector(".jb-job-categories");e&&e.addEventListener("click",(t=>{t.target!==e&&(t.preventDefault(),t.stopPropagation())}))}))}))})).observe(document,{attributes:!1,childList:!0,characterData:!1,subtree:!0})})),(0,t.registerBlockType)("jb-block/jb-jobs-categories-list",{edit:function(e){const t=(0,n.useBlockProps)();return(0,c.jsx)("div",{...t,children:(0,c.jsx)(r(),{block:"jb-block/jb-jobs-categories-list"})})},save:function(){return null}})})();
