+++ includes/blocks/jb-jobs-dashboard/src/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-server-side-render'), 'version' => 'f87d6fe27c4e440862e8');
