+++ includes/blocks/jb-jobs-dashboard/src/index.js
@@ -1,59 +1,1 @@
-jQuery(window).on( 'load', function($) {
-	var observer = new MutationObserver(function(mutations) {
-		mutations.forEach(function(mutation) {
-
-			jQuery(mutation.addedNodes).find('.jb-job-dashboard').each(function() {
-				wp.JB.jobs_dashboard.objects.wrapper = jQuery('.jb-job-dashboard');
-				if ( wp.JB.jobs_dashboard.objects.wrapper.length ) {
-					wp.JB.jobs_dashboard.ajax();
-				}
-				const wrapper = document.querySelector('.jb-job-dashboard');
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-
-			jQuery(mutation.addedNodes).find('.jb').each(function() {
-				jb_responsive();
-				const wrapper = document.querySelector('.jb-job-dashboard');
-
-				if (wrapper) {
-					wrapper.addEventListener('click', (event) => {
-						if (event.target !== wrapper) {
-							event.preventDefault();
-							event.stopPropagation();
-						}
-					});
-				}
-			});
-		});
-	});
-
-	observer.observe(document, {attributes: false, childList: true, characterData: false, subtree:true});
-});
-
-import { registerBlockType } from '@wordpress/blocks';
-import ServerSideRender from '@wordpress/server-side-render';
-import { useBlockProps } from '@wordpress/block-editor';
-
-registerBlockType('jb-block/jb-jobs-dashboard', {
-	edit: function (props) {
-		const blockProps = useBlockProps();
-
-		return (
-			<div {...blockProps}>
-				<ServerSideRender block="jb-block/jb-jobs-dashboard" />
-			</div>
-		);
-	},
-
-	save: function () {
-		return null;
-	}
-});
+(()=>{"use strict";var e={n:o=>{var r=o&&o.__esModule?()=>o.default:()=>o;return e.d(r,{a:r}),r},d:(o,r)=>{for(var t in r)e.o(r,t)&&!e.o(o,t)&&Object.defineProperty(o,t,{enumerable:!0,get:r[t]})},o:(e,o)=>Object.prototype.hasOwnProperty.call(e,o)};const o=window.wp.blocks,r=window.wp.serverSideRender;var t=e.n(r);const a=window.wp.blockEditor,n=window.ReactJSXRuntime;jQuery(window).on("load",(function(e){new MutationObserver((function(e){e.forEach((function(e){jQuery(e.addedNodes).find(".jb-job-dashboard").each((function(){wp.JB.jobs_dashboard.objects.wrapper=jQuery(".jb-job-dashboard"),wp.JB.jobs_dashboard.objects.wrapper.length&&wp.JB.jobs_dashboard.ajax();const e=document.querySelector(".jb-job-dashboard");e&&e.addEventListener("click",(o=>{o.target!==e&&(o.preventDefault(),o.stopPropagation())}))})),jQuery(e.addedNodes).find(".jb").each((function(){jb_responsive();const e=document.querySelector(".jb-job-dashboard");e&&e.addEventListener("click",(o=>{o.target!==e&&(o.preventDefault(),o.stopPropagation())}))}))}))})).observe(document,{attributes:!1,childList:!0,characterData:!1,subtree:!0})})),(0,o.registerBlockType)("jb-block/jb-jobs-dashboard",{edit:function(e){const o=(0,a.useBlockProps)();return(0,n.jsx)("div",{...o,children:(0,n.jsx)(t(),{block:"jb-block/jb-jobs-dashboard"})})},save:function(){return null}})})();
