+++ includes/blocks/jb-jobs-list/src/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-data', 'wp-element', 'wp-server-side-render'), 'version' => '5d6b917dbc001d9dc40a');
