+++ includes/blocks/jb-recent-jobs/src/index.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react-jsx-runtime', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-data', 'wp-element', 'wp-server-side-render'), 'version' => '3bca1ae6e646cd61bc9b');
