+++ phpcs.xml
@@ -16,6 +16,8 @@
 	<exclude-pattern>documentor/*</exclude-pattern>
 	<exclude-pattern>\.idea/*</exclude-pattern>
 	<exclude-pattern>includes/blocks/*</exclude-pattern>
+	<exclude-pattern>blocks-src/*</exclude-pattern>
+	<exclude-pattern>build/*</exclude-pattern>
 
 	<rule ref="WordPress-Extra"/>
